feat(server): KingSystem Phase 1, +monsummon/+add_xp GM commands, exp overflow fix
Add CKingSystem class with WIZ_KING router, in-memory per-nation state, KING_TAX/KING_EVENT/KING_NPC handlers, and event tick. King gets Rank=1/Title=1 and the king cape (Cloak.tbl id 99) is broadcast in WIZ_MYINFO + GetUserInfo regardless of clan.

New AG_NPC_SUMMON_REQ (=14) opcode lets +monsummon spawn a CNpc on AIServer, register it on Ebenezer via NPC_INFO_ALL, then broadcast via AG_NPC_INFO. Serial range 20000+ keeps wire id (sNid+NPC_BAND) within positive int16_t.

AG_USER_EXP wire format widened from int16_t to int32_t on both sides — high-level mob kills were generating exp values that overflowed and were rejected with 'invalid exp or loyalty amount granted'.

GM commands: +set_king, +king_event, +king_tax, +king_status, +add_xp / +add_exp.

+++ src/Server/AIServer/User.cpp
@@ -450,8 +450,10 @@ void CUser::SendExp(int iExp, int iLoyalty, int /*tType*/)
 
 	SetByte(buff, AG_USER_EXP, sendIndex);
 	SetShort(buff, m_iUserId, sendIndex);
-	SetShort(buff, iExp, sendIndex);
-	SetShort(buff, iLoyalty, sendIndex);
+	// Use 32-bit fields: int16_t overflowed for high-level monsters and Ebenezer
+	// would reject the packet ("invalid exp or loyalty amount granted").
+	SetInt(buff, iExp, sendIndex);
+	SetInt(buff, iLoyalty, sendIndex);
 
 	//TRACE(_T("$$ User - SendExp : %hs, exp=%d, loyalty=%d $$\n"), m_strUserID, iExp, iLoyalty);
